Добавляет слайс комментариев
